create bin2rom

+++ src/headers/cpu6502.h
@@ -8,7 +8,7 @@
 typedef struct {
     uint8_t ir;// Instruction Register
     uint8_t tcu;// Timing Control Unit
-    uint8_t pc;
+    uint16_t pc;
     uint8_t s;
     uint8_t x;
     uint8_t y;
